requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect2==0.4.1
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-appconf==1.0.4
django-auditlog==1.0a1
django-braces==1.14.0
django-cpf==0.1.0
django-crispy-forms==1.9.1
django-global-permissions==0.2.4
django-jsonfield==1.4.0
django-mail-templated==2.6.5
django-querysetsequence==0.12
django-redis==4.11.0
django-sequences==2.5
django-smtp-ssl==1.0
django-sortedm2m==3.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
easy-mask==1.3.4
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
html5lib==1.1
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
openpyxl==3.0.4
packaging==20.4
pandas==1.2.1
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2==2.8.5
pycparser==2.20
pycpfcnpj==1.5.1
Pygments==2.6.1
pyodbc==4.0.30
pyparsing==2.4.7
Pyphen==0.10.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.1
python-http-client==3.3.1
pytz==2019.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
redis==3.5.2
reportlab==3.5.59
requests==2.23.0
selenium==3.141.0
Send2Trash==1.5.0
sendgrid==6.4.7
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.0
starkbank-ecdsa==1.1.0
terminado==0.8.3
testpath==0.4.4
tinycss2==1.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
waitress==1.4.4
wcwidth==0.2.3
WeasyPrint==52.2
webencodings==0.5.1
Werkzeug==1.0.1
wfastcgi==3.0.0
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
